rename: study dto에서 goalAchieve와 attendance관련 dto 위치 변경
