+++ src/main/java/com/grepp/spring/app/controller/api/StudyController.java
@@ -2,7 +2,7 @@ package com.grepp.spring.app.controller.api;
 
 import com.grepp.spring.app.model.applicant.dto.request.ApplicationRequest;
 import com.grepp.spring.app.model.applicant.dto.request.ApplicationResultRequest;
-import com.grepp.spring.app.model.study.dto.response.CheckGoalResponse;
+import com.grepp.spring.app.model.studyMmeber.dto.response.CheckGoalResponse;
 import com.grepp.spring.app.model.study.dto.request.NotificationUpdateRequest;
 import com.grepp.spring.app.model.study.dto.request.StudyCreationRequest;
 import com.grepp.spring.app.model.study.dto.request.StudySearchRequest;
@@ -17,8 +17,8 @@ import com.grepp.spring.app.model.study.code.Status;
 import com.grepp.spring.app.model.study.dto.response.StudyCreationResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyInfoResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyListResponse;
-import com.grepp.spring.app.model.study.dto.response.WeeklyAttendanceResponse;
-import com.grepp.spring.app.model.study.dto.response.WeeklyGoalStatusResponse;
+import com.grepp.spring.app.model.studyMmeber.dto.response.WeeklyAttendanceResponse;
+import com.grepp.spring.app.model.studyMmeber.dto.response.WeeklyGoalStatusResponse;
 import com.grepp.spring.app.model.study.dto.response.GoalsResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyNoticeResponse;
 import com.grepp.spring.app.model.applicant.service.ApplicantService;
