+++ src/main/java/com/grepp/spring/app/model/studyMmeber/repository/goalAchievement/GoalRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.grepp.spring.app.model.studyMmeber.repository.goalAchievement;
 
-import com.grepp.spring.app.model.study.dto.response.CheckGoalResponse;
-import com.grepp.spring.app.model.study.dto.response.WeeklyAchievementCount;
+import com.grepp.spring.app.model.studyMmeber.dto.response.CheckGoalResponse;
+import com.grepp.spring.app.model.studyMmeber.dto.response.WeeklyAchievementCount;
 import com.grepp.spring.app.model.study.entity.StudyGoal;
 import com.grepp.spring.app.model.study.dto.response.GoalsResponse;
 import java.time.LocalDate;
